Fix comparison warning in test_buffer

+++ src/tests/test_buffer.c
@@ -70,14 +70,14 @@ static void buffer_create_info_destroy_test() {
 	host_ptr = ccl_memobj_get_info_scalar(
 		b, CL_MEM_HOST_PTR, void*, &err);
 	g_assert((err == NULL) || (err->code == CCL_ERROR_INFO_UNAVAILABLE_OCL));
-	g_assert_cmphex(host_ptr, ==, NULL);
-    g_clear_error(&err);
+	g_assert_cmphex((gulong) host_ptr, ==, (gulong) NULL);
+	g_clear_error(&err);
 
 	cl_context context;
 	context = ccl_memobj_get_info_scalar(
 		b, CL_MEM_CONTEXT, cl_context, &err);
 	g_assert_no_error(err);
-	g_assert_cmphex(context, ==, ccl_context_unwrap(ctx));
+	g_assert_cmphex((gulong) context, ==, (gulong) ccl_context_unwrap(ctx));
 
 	/* Destroy stuff. */
 	ccl_buffer_destroy(b);
